// src/types/article.ts
export interface User {
  id: number;
  username: string;
  email: string;
}

export interface Comment {
  id: number;
  content: string;
}

export interface Category {
  id: number;
  name: string;
}

export interface Article {
  id: number;
  documentId: string;
  title: string;
  description: string;
  cover_image_url: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  user: User;
  comments: Comment[];
  category: Category | null;
}
export interface Pagination {
  page: number;
  pageSize: number;
  pageCount: number;
  total: number;
}